src/backend/pyproject.toml
[project]
name = "nlp-projet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "decord>=0.6.0",
    "einops>=0.8.2",
    "opencv-python>=4.12.0.88",
    "pillow",
    "ruff>=0.14.9",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.52.1,<5.0.0",
]

[project.scripts]
stream = "streaming.VideoStream:main"
record = "streaming.RegisterStream:main"
inference = "streaming.inference_multi:main"
app = "main:main"
